parser/requirements.txt
lxml
debugpy
python-gitlab
Jinja2
